visual_intelligent_assistant/requirements.txt
streamlit
openai
pandas
pillow
PyPDF2
SpeechRecognition